An exported struct field governs how often ReaderCheck should be called
The default value causes MDBStore to call ReaderCheck only once at
initialization, which seems like the correct default behavior.  Repeated
calls to ReaderCheck is only a benefit for systems requiring
multi-process database access.  Because github.com/hashicorp/raft
primarily is backed by bolt, multi-process access should not be a
requirement and need not be anticipated in the MDBStore by default.